+++ networkml/helpers/gzipio.py
